// src/enums/compatibility/EIDColor.ts
/** Shortcuts for setting the Color of EID Text. */
export enum EIDColorShortcut {
  RED = "red_del",

  /** HAPPY99. */
  HAPPY_YELLOW = "happy_yellow_del",
  HAPPY_YELLOW_DARKER = "happy_yellow_darker_del",
  HAPPY_YELLOW_DARKEST = "happy_yellow_darkest_del",

  LOVE_PINK = "love_pink_del",
  WORM_TURQUOISE = "worm_turquoise_del",
  ANGRY_RED = "angry_red_del",
  DEATH_BLACK = "death_black_del",
}